requirements.txt
sqlalchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
requests>=2.31.0